code/pyproject.toml
[project]
name = "nano-banana-desktop"
version = "0.1.0"
description = "Desktop image editing utility using Google's Gemini Flash 2.5 Images (Nano Banana)"
authors = [
    {name = "Daniel Rosehill", email = "public@danielrosehill.com"}
]
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.4.0",
    "pillow>=10.0.0",
    "pyside6>=6.6.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["nano_banana", "nano_banana.ui", "nano_banana.core", "nano_banana.utils"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
